Move partial migration classes to the demo module.
