+++ demo-simple/src/main/java/com/novoda/downloadmanager/demo/migration/PartialDownloadBatchesExtractor.java
@@ -1,6 +1,14 @@
-package com.novoda.downloadmanager;
+package com.novoda.downloadmanager.demo.migration;
 
 import android.database.Cursor;
+
+import com.novoda.downloadmanager.Batch;
+import com.novoda.downloadmanager.BatchBuilder;
+import com.novoda.downloadmanager.DownloadBatchId;
+import com.novoda.downloadmanager.DownloadBatchIdCreator;
+import com.novoda.downloadmanager.DownloadFileIdCreator;
+import com.novoda.downloadmanager.MigrationStoragePathSanitizer;
+import com.novoda.downloadmanager.SqlDatabaseWrapper;
 
 import java.util.ArrayList;
 import java.util.HashSet;
